fix: finish compilation

+++ src/JualIn.App.Mobile.Presentation/Modules/Sales/ViewModels/SaleProductViewModel.cs
@@ -12,7 +12,7 @@ public partial class SaleProductViewModel(Product @entity) : ObservableObject
     [ObservableProperty]
     [NotifyPropertyChangedFor(nameof(OnCartTotalPrice))]
     [NotifyPropertyChangedFor(nameof(IsSelected))]
-    private double _onCartQuantity;
+    private int _onCartQuantity;
 
     public double OnCartTotalPrice => Entity.Price * OnCartQuantity;
